Edits based on Management review
A variety of edits based on management review, including:

* Moved more of the explanatory notes from the middle of the document
to the top.

* Removed word “initial” from before charter (2 instances)

* New subsection header on “benefits” in the goals section

* Changed “wallet” to “digital wallet” globally to use one term
constantly throughout the charter.

* Removed observation about “increased competition” from one of the
benefits.

* Rewrote the front of the scope section. Moved non-essential text to
the FAQ. Crisper statements about what is in scope and what is not.

* Eliminated the “out of scope” section since, after comments, the only
sentence left was that the group would not define a payment scheme.
Moved that sentence to earlier in the document, with a link to the
definition of “payment scheme” in the use cases.

* Moved flow diagram to the FAQ as it was perceived to lengthen the
charter and was non-normative.

* Rewrote the security section to make clearer that at one level
(message verification) we need to ensure these APIs are secure. But
that at another level (user auth) we are leaving that to the schemes
and simply monitoring what other groups are doing (that might be
adopted by the schemes).

* NEW subsection with a statement about regulatory: the standards
should not prevent schemes from fulfilling their regulatory obligations.

* Merged the two specifications into a single specification since the
vocabularies will not be used “on their own” but simply be part of the
API. The vocabulary text is still part of the description, just one
spec instead of two.

* Changed “WebIDL API” to “JavaScript API”.

* Changed the description of the two APIs to be clearer about their
roles (app to user agent, and user agent to server-side wallet).

* Changed REST to server-side.

* The user agent to server-side solution MAY be HTTP based but
management review suggests that some groups are taking alternative
approaches and so the WG should have the flexibility to choose the best
approach.

* Removed “1.0” designation from specifications. (Apparently that is
not really done anymore.)

* Added liaisons to Web Apps and WebAppSec

* Clarified that we expect 10 active participants (instead of just
“active participants”)

* Mention role of EU support in creation of this charter

* Fixed URI in address.